combine integration tests
